Splitting clients out into separate .py file and creating subclass structure for different streams.
